+++ pkg/binder/binding/default_binder_test.go
@@ -25,11 +25,35 @@ import (
 	"github.com/NVIDIA/KAI-scheduler/pkg/binder/test_utils"
 )
 
+const (
+	gpuSharingConfigMapAnnotation = "runai/shared-gpu-configmap"
+)
+
 func TestBind(t *testing.T) {
 	pod := &v1.Pod{
 		ObjectMeta: metav1.ObjectMeta{
 			Namespace: "my-ns",
 			Name:      "my-pod",
+			Annotations: map[string]string{
+				gpuSharingConfigMapAnnotation: "my-configmap-shared-gpu",
+			},
+		},
+		Spec: v1.PodSpec{
+			Containers: []v1.Container{
+				{},
+			},
+			Volumes: []v1.Volume{
+				{
+					Name: "my-configmap-vol",
+					VolumeSource: v1.VolumeSource{
+						ConfigMap: &v1.ConfigMapVolumeSource{
+							LocalObjectReference: v1.LocalObjectReference{
+								Name: "my-configmap-shared-gpu-0",
+							},
+						},
+					},
+				},
+			},
 		},
 	}
 	kubeObjects := []runtime.Object{
@@ -74,6 +98,9 @@ func TestBindApplyResourceReceivedType(t *testing.T) {
 		ObjectMeta: metav1.ObjectMeta{
 			Namespace: "my-ns",
 			Name:      "my-pod",
+			Annotations: map[string]string{
+				gpuSharingConfigMapAnnotation: "my-config",
+			},
 		},
 		Spec: v1.PodSpec{
 			Containers: []v1.Container{
@@ -85,7 +112,7 @@ func TestBindApplyResourceReceivedType(t *testing.T) {
 								ConfigMapKeyRef: &v1.ConfigMapKeySelector{
 									Key: common.VisibleDevices,
 									LocalObjectReference: v1.LocalObjectReference{
-										Name: "my-config",
+										Name: "my-config-0",
 									},
 								},
 							},
@@ -96,9 +123,21 @@ func TestBindApplyResourceReceivedType(t *testing.T) {
 								ConfigMapKeyRef: &v1.ConfigMapKeySelector{
 									Key: common.NumOfGpusEnvVar,
 									LocalObjectReference: v1.LocalObjectReference{
-										Name: "my-config",
+										Name: "my-config-0",
 									},
 								},
+							},
+						},
+					},
+				},
+			},
+			Volumes: []v1.Volume{
+				{
+					Name: "my-configmap-vol",
+					VolumeSource: v1.VolumeSource{
+						ConfigMap: &v1.ConfigMapVolumeSource{
+							LocalObjectReference: v1.LocalObjectReference{
+								Name: "my-config-0",
 							},
 						},
 					},