+++ pkg/binder/plugins/gpusharing/gpu_sharing.go
@@ -22,9 +22,8 @@ import (
 )
 
 const (
-	fractionContainerIndex        = 0
-	gpuSharingConfigMapAnnotation = "runai/shared-gpu-configmap"
-	CdiDeviceNameBase             = "k8s.device-plugin.nvidia.com/gpu=%s"
+	fractionContainerIndex = 0
+	CdiDeviceNameBase      = "k8s.device-plugin.nvidia.com/gpu=%s"
 )
 
 type GPUSharing struct {
@@ -64,18 +63,20 @@ func (p *GPUSharing) Mutate(pod *v1.Pod) error {
 		return nil
 	}
 
-	capabilitiesConfigMapName := gpusharingconfigmap.SetGpuCapabilitiesConfigMapName(pod, fractionContainerIndex,
-		gpusharingconfigmap.RegularContainer)
-	directEnvVarsMapName, err := gpusharingconfigmap.ExtractDirectEnvVarsConfigMapName(pod, fractionContainerIndex,
-		gpusharingconfigmap.RegularContainer)
+	containerRef := &gpusharingconfigmap.PodContainerRef{
+		Container: &pod.Spec.Containers[fractionContainerIndex],
+		Index:     fractionContainerIndex,
+		Type:      gpusharingconfigmap.RegularContainer,
+	}
+	capabilitiesConfigMapName := gpusharingconfigmap.SetGpuCapabilitiesConfigMapName(pod, containerRef)
+	directEnvVarsMapName, err := gpusharingconfigmap.ExtractDirectEnvVarsConfigMapName(pod, containerRef)
 	if err != nil {
 		return err
 	}
 
-	fractionContainer := &pod.Spec.Containers[fractionContainerIndex]
-	common.AddVisibleDevicesEnvVars(fractionContainer, capabilitiesConfigMapName)
+	common.AddVisibleDevicesEnvVars(containerRef.Container, capabilitiesConfigMapName)
 	common.SetConfigMapVolume(pod, capabilitiesConfigMapName)
-	common.AddDirectEnvVarsConfigMapSource(fractionContainer, directEnvVarsMapName)
+	common.AddDirectEnvVarsConfigMapSource(containerRef.Container, directEnvVarsMapName)
 
 	return nil
 }
@@ -94,53 +95,44 @@ func (p *GPUSharing) PreBind(
 		}
 	}
 
-	err, legacyPod := p.createCapabilitiesConfigMapIfMissing(ctx, pod)
+	containerRef := &gpusharingconfigmap.PodContainerRef{
+		Container: &pod.Spec.Containers[fractionContainerIndex],
+		Index:     fractionContainerIndex,
+		Type:      gpusharingconfigmap.RegularContainer,
+	}
+	err := p.createCapabilitiesConfigMapIfMissing(ctx, pod, containerRef)
 	if err != nil {
 		return fmt.Errorf("failed to create capabilities configmap: %w", err)
 	}
 
-	if !legacyPod {
-		err = p.createDirectEnvMapIfMissing(ctx, pod)
-		if err != nil {
-			return fmt.Errorf("failed to create env configmap: %w", err)
-		}
+	err = p.createDirectEnvMapIfMissing(ctx, pod, containerRef)
+	if err != nil {
+		return fmt.Errorf("failed to create env configmap: %w", err)
 	}
 
-	fractionContainer := &pod.Spec.Containers[fractionContainerIndex]
 	nVisibleDevicesStr := strings.Join(reservedGPUIds, ",")
-	err = common.SetNvidiaVisibleDevices(ctx, p.kubeClient, pod, fractionContainer, nVisibleDevicesStr)
+	err = common.SetNvidiaVisibleDevices(ctx, p.kubeClient, pod, containerRef, nVisibleDevicesStr)
 	if err != nil {
 		return err
 	}
 
 	numOfGPUDevices := fmt.Sprintf("%v", bindRequest.Spec.ReceivedGPU.Portion)
-	return common.SetNumOfGPUDevices(ctx, p.kubeClient, pod, fractionContainer, numOfGPUDevices)
+	return common.SetNumOfGPUDevices(ctx, p.kubeClient, pod, containerRef, numOfGPUDevices)
 }
 
-func (p *GPUSharing) createCapabilitiesConfigMapIfMissing(ctx context.Context, pod *v1.Pod) (error, bool) {
-	legacyPod := false
-	var capabilitiesConfigMapName string
-	var err error
-	_, found := pod.Annotations[gpuSharingConfigMapAnnotation]
-	if found {
-		capabilitiesConfigMapName, err = gpusharingconfigmap.ExtractCapabilitiesConfigMapName(pod,
-			fractionContainerIndex, gpusharingconfigmap.RegularContainer)
-	} else {
-		legacyPod = true
-		fractionContainer := &pod.Spec.Containers[fractionContainerIndex]
-		capabilitiesConfigMapName, err = common.GetConfigMapName(pod, fractionContainer)
-	}
-
+func (p *GPUSharing) createCapabilitiesConfigMapIfMissing(ctx context.Context, pod *v1.Pod,
+	containerRef *gpusharingconfigmap.PodContainerRef) error {
+	capabilitiesConfigMapName, err := gpusharingconfigmap.ExtractCapabilitiesConfigMapName(pod, containerRef)
 	if err != nil {
-		return fmt.Errorf("failed to get capabilities configmap name: %w", err), false
+		return fmt.Errorf("failed to get capabilities configmap name: %w", err)
 	}
 	err = gpusharingconfigmap.UpsertJobConfigMap(ctx, p.kubeClient, pod, capabilitiesConfigMapName, map[string]string{})
-	return err, legacyPod
+	return err
 }
 
-func (p *GPUSharing) createDirectEnvMapIfMissing(ctx context.Context, pod *v1.Pod) error {
-	directEnvVarsMapName, err := gpusharingconfigmap.ExtractDirectEnvVarsConfigMapName(pod,
-		fractionContainerIndex, gpusharingconfigmap.RegularContainer)
+func (p *GPUSharing) createDirectEnvMapIfMissing(ctx context.Context, pod *v1.Pod,
+	containerRef *gpusharingconfigmap.PodContainerRef) error {
+	directEnvVarsMapName, err := gpusharingconfigmap.ExtractDirectEnvVarsConfigMapName(pod, containerRef)
 	if err != nil {
 		return err
 	}